add p2p submit for full node miner, add rpc submit for just remote miner
