+++ problem/problem_test.go
@@ -3,8 +3,6 @@ package problem
 import (
 	"bytes"
 	"crypto/sha256"
-	"errors"
-	"fmt"
 	"github.com/Evanesco-Labs/miner/log"
 	"github.com/consensys/gnark-crypto/ecc"
 	"github.com/consensys/gnark-crypto/ecc/bn254/fr"
@@ -12,22 +10,11 @@ import (
 	"github.com/consensys/gnark/backend"
 	"github.com/consensys/gnark/backend/groth16"
 	"github.com/consensys/gnark/frontend"
-	"github.com/ethereum/go-ethereum/consensus/ethash"
-	"github.com/ethereum/go-ethereum/core"
-	"github.com/ethereum/go-ethereum/core/rawdb"
-	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/params"
 	"github.com/stretchr/testify/assert"
 	"os"
 	"runtime"
 	"testing"
 	"time"
-)
-
-var (
-	PkPath   = "../provekey.txt"
-	VkPath   = "../verifykey.txt"
-	R1csPath = "../r1cs.txt"
 )
 
 func TestNewProvingKey(t *testing.T) {
@@ -129,69 +116,6 @@ func TestCircuit(t *testing.T) {
 	}
 }
 
-func GenerateTestChain(cnt int) (*core.HeaderChain, error) {
-	testdb := rawdb.NewMemoryDatabase()
-	gspec := &core.Genesis{Config: params.TestChainConfig}
-	genesis := gspec.MustCommit(testdb)
-	hc, err := core.NewHeaderChain(testdb, params.TestChainConfig, ethash.NewFaker(), func() bool { return false })
-	if err != nil {
-		return nil, err
-	}
-	hc.SetGenesis(genesis.Header())
-	blocks, _ := core.GenerateChain(params.TestChainConfig, genesis, ethash.NewFaker(), testdb, 100, nil)
-	headerChain := make([]*types.Header, 0)
-	for _, block := range blocks {
-		headerChain = append(headerChain, block.Header())
-	}
-
-	_, err = hc.InsertHeaderChain(headerChain, time.Now())
-	if err != nil {
-		return nil, err
-	}
-	return hc, nil
-}
-
-func TestNewProverVerifier(t *testing.T) {
-	hc, err := GenerateTestChain(100)
-	if err != nil {
-		panic(err)
-	}
-	runtime.GOMAXPROCS(1)
-	t1 := time.Now()
-	prover, err := NewProblemProver(PkPath)
-	t2 := time.Now()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	msg := "test evanesco"
-	hash := sha256.New()
-	hash.Write([]byte(msg))
-	preimage := hash.Sum(nil)
-
-	mimcHash, proof := prover.Prove(preimage)
-	t3 := time.Now()
-	getHeader := func(height uint64) (*types.Header, error) {
-		header := hc.GetHeaderByNumber(height)
-		if header == nil {
-			return nil, errors.New("no header")
-		}
-		return header, nil
-	}
-	verifier, err := NewProblemVerifier(VkPath, 100, 80, getHeader)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t4 := time.Now()
-	result := verifier.VerifyZKP(preimage, mimcHash, proof)
-	t5 := time.Now()
-	fmt.Println("new prover:", t2.Sub(t1).String())
-	fmt.Println("prove:", t3.Sub(t2).String())
-	fmt.Println("new verifier:", t4.Sub(t3).String())
-	fmt.Println("verify:", t5.Sub(t4).String())
-	assert.Equal(t, true, result)
-}
-
 //TestSetupZKP checks setup randomness
 func TestSetupZKP(t *testing.T) {
 	r1cs := CompileCircuit()
@@ -211,8 +135,4 @@ func TestSetupZKP(t *testing.T) {
 	vk1.WriteTo(buf1)
 	vk2.WriteTo(buf2)
 	assert.Equal(t, false, bytes.Equal(buf1.Bytes(), buf2.Bytes()))
-}
-
-func TestLottery_Serialize(t *testing.T) {
-
 }
